+++ experiments/firmware/pigss/boxer/src/tca954xa.c
@@ -153,7 +153,7 @@ int16_t tca9548a_read(uint8_t slave_address) {
 }
 
 int8_t tca9548a_write(uint8_t slave_address, uint8_t command_byte) {
-  int16_t retval = 0;
+  int8_t retval = 0;
   // Keep track of the number of times we've tried to write
   uint8_t tries = 0;
   // Slave address + R/W bit
